refactor: delete hoc directory
